+++ src/api/dto/settings.dto.ts
@@ -6,4 +6,5 @@ export class SettingsDto {
   readMessages?: boolean;
   readStatus?: boolean;
   syncFullHistory?: boolean;
+  wavoipToken?: string;
 }