+++ src/api/integrations/event/pusher/pusher.controller.ts
@@ -120,7 +120,11 @@ export class PusherController extends EventController implements EventController
     sender,
     apiKey,
     local,
+    integration,
   }: EmitData): Promise<void> {
+    if (integration && !integration.includes('pusher')) {
+      return;
+    }
     if (!this.status) {
       return;
     }