+++ src/api/integrations/event/rabbitmq/rabbitmq.controller.ts
@@ -73,7 +73,12 @@ export class RabbitmqController extends EventController implements EventControll
     dateTime,
     sender,
     apiKey,
+    integration,
   }: EmitData): Promise<void> {
+    if (integration && !integration.includes('rabbitmq')) {
+      return;
+    }
+
     if (!this.status) {
       return;
     }
@@ -82,6 +87,7 @@ export class RabbitmqController extends EventController implements EventControll
     const rabbitmqLocal = instanceRabbitmq?.events;
     const rabbitmqGlobal = configService.get<Rabbitmq>('RABBITMQ').GLOBAL_ENABLED;
     const rabbitmqEvents = configService.get<Rabbitmq>('RABBITMQ').EVENTS;
+    const prefixKey = configService.get<Rabbitmq>('RABBITMQ').PREFIX_KEY;
     const rabbitmqExchangeName = configService.get<Rabbitmq>('RABBITMQ').EXCHANGE_NAME;
     const we = event.replace(/[.-]/gm, '_').toUpperCase();
     const logEnabled = configService.get<Log>('LOG').LEVEL.includes('WEBHOOKS');
@@ -154,7 +160,9 @@ export class RabbitmqController extends EventController implements EventControll
             autoDelete: false,
           });
 
-          const queueName = event;
+          const queueName = prefixKey
+            ? `${prefixKey}.${event.replace(/_/g, '.').toLowerCase()}`
+            : event.replace(/_/g, '.').toLowerCase();
 
           await this.amqpChannel.assertQueue(queueName, {
             durable: true,
@@ -190,6 +198,7 @@ export class RabbitmqController extends EventController implements EventControll
 
     const rabbitmqExchangeName = configService.get<Rabbitmq>('RABBITMQ').EXCHANGE_NAME;
     const events = configService.get<Rabbitmq>('RABBITMQ').EVENTS;
+    const prefixKey = configService.get<Rabbitmq>('RABBITMQ').PREFIX_KEY;
 
     if (!events) {
       this.logger.warn('No events to initialize on AMQP');
@@ -202,7 +211,10 @@ export class RabbitmqController extends EventController implements EventControll
     eventKeys.forEach((event) => {
       if (events[event] === false) return;
 
-      const queueName = `${event.replace(/_/g, '.').toLowerCase()}`;
+      const queueName =
+        prefixKey !== ''
+          ? `${prefixKey}.${event.replace(/_/g, '.').toLowerCase()}`
+          : `${event.replace(/_/g, '.').toLowerCase()}`;
       const exchangeName = rabbitmqExchangeName;
 
       this.amqpChannel.assertExchange(exchangeName, 'topic', {