+++ src/api/integrations/event/sqs/sqs.controller.ts
@@ -54,7 +54,12 @@ export class SqsController extends EventController implements EventControllerInt
     dateTime,
     sender,
     apiKey,
+    integration,
   }: EmitData): Promise<void> {
+    if (integration && !integration.includes('sqs')) {
+      return;
+    }
+
     if (!this.status) {
       return;
     }